separate logic and data#20
